refactor: Revert to first MVP (#17)
### Description

After further analysis of the requirements required for the next MVP iteration, the best way to move forward is to work on the previous version of MVP instead of heading straight into the scalability concept. As such, I have included this commit to revert the previous work-in-progress changes made for scalability purpose for quicker development and prototyping. 

* Revert all folders except for app folder

* Add app folder

* Add linting and formatting changes made previously